+++ stubs/DecodeLabs/Guidance.php
@@ -38,26 +38,50 @@ class Guidance implements Proxy
     public static function createVoid(): Ref2 {
         return static::$instance->createVoid();
     }
+    public static function createVoidString(): string {
+        return static::$instance->createVoidString();
+    }
     public static function createV1(string|int|null $node = NULL, ?int $clockSeq = NULL): Ref2 {
         return static::$instance->createV1(...func_get_args());
+    }
+    public static function createV1String(string|int|null $node = NULL, ?int $clockSeq = NULL): string {
+        return static::$instance->createV1String(...func_get_args());
     }
     public static function createV3(string $name, ?string $namespace = NULL): Ref2 {
         return static::$instance->createV3(...func_get_args());
     }
+    public static function createV3String(string $name, ?string $namespace = NULL): string {
+        return static::$instance->createV3String(...func_get_args());
+    }
     public static function createV4(): Ref2 {
         return static::$instance->createV4();
+    }
+    public static function createV4String(): string {
+        return static::$instance->createV4String();
     }
     public static function createV4Comb(): Ref2 {
         return static::$instance->createV4Comb();
     }
+    public static function createV4CombString(): string {
+        return static::$instance->createV4CombString();
+    }
     public static function createV5(string $name, ?string $namespace = NULL): Ref2 {
         return static::$instance->createV5(...func_get_args());
+    }
+    public static function createV5String(string $name, ?string $namespace = NULL): string {
+        return static::$instance->createV5String(...func_get_args());
     }
     public static function createV6(string|int|null $node = NULL, ?int $clockSeq = NULL): Ref2 {
         return static::$instance->createV6(...func_get_args());
     }
+    public static function createV6String(string|int|null $node = NULL, ?int $clockSeq = NULL): string {
+        return static::$instance->createV6String(...func_get_args());
+    }
     public static function createV7(?Ref3 $date = NULL): Ref2 {
         return static::$instance->createV7(...func_get_args());
+    }
+    public static function createV7String(?Ref3 $date = NULL): string {
+        return static::$instance->createV7String(...func_get_args());
     }
     public static function isValid(Ref4|Ref2|string|null $uuid): bool {
         return static::$instance->isValid(...func_get_args());