+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added create*String() methods
+
 ## v0.1.1 (2023-11-14)
 * Added full-length check to fromShortString()